Re-organize some classes
